refactor(core): Rename parse_hal and validate_hal
Partly this is because go discourages package names including underscores
(excluding _test packages). In the case of parse_hal, rename it transform
as it's primarily used to transform protos.
