+++ pkg/validate/validate.go
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package validate_hal
+package validate
 
 import (
 	"errors"
