+++ pkg/validate/validate_test.go
@@ -13,19 +13,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package validate_hal_test
+package validate_test
 
 import (
 	"testing"
 
 	"github.com/spinnaker/kleat/api/client/cloudprovider"
 	"github.com/spinnaker/kleat/api/client/config"
-	"github.com/spinnaker/kleat/pkg/validate_hal"
+	"github.com/spinnaker/kleat/pkg/validate"
 )
 
 func TestEmptyHalConfig(t *testing.T) {
 	h := &config.Hal{}
-	err := validate_hal.ValidateHalConfig(h)
+	err := validate.ValidateHalConfig(h)
 	if err != nil {
 		t.Errorf("Expected no validation failures, got %s", err)
 	}
@@ -35,7 +35,7 @@ func TestEmptyProviders(t *testing.T) {
 	h := &config.Hal{
 		Providers: &cloudprovider.Providers{},
 	}
-	err := validate_hal.ValidateHalConfig(h)
+	err := validate.ValidateHalConfig(h)
 	if err != nil {
 		t.Errorf("Expected no validation failures, got %s", err)
 	}
@@ -47,7 +47,7 @@ func TestEmptyKubernetes(t *testing.T) {
 			Kubernetes: &cloudprovider.Kubernetes{},
 		},
 	}
-	err := validate_hal.ValidateHalConfig(h)
+	err := validate.ValidateHalConfig(h)
 	if err != nil {
 		t.Errorf("Expected no validation failures, got %s", err)
 	}
@@ -62,7 +62,7 @@ func TestNoKubernetesAccounts(t *testing.T) {
 			},
 		},
 	}
-	err := validate_hal.ValidateHalConfig(h)
+	err := validate.ValidateHalConfig(h)
 	if err != nil {
 		t.Errorf("Expected no validation failures, got %s", err)
 	}
@@ -81,7 +81,7 @@ func TestKuberntesAccountWithNoOmitKinds(t *testing.T) {
 			},
 		},
 	}
-	err := validate_hal.ValidateHalConfig(h)
+	err := validate.ValidateHalConfig(h)
 	if err != nil {
 		t.Errorf("Expected no validation failures, got %s", err)
 	}
@@ -101,7 +101,7 @@ func TestKuberntesAccountWithEmptyOmitKinds(t *testing.T) {
 			},
 		},
 	}
-	err := validate_hal.ValidateHalConfig(h)
+	err := validate.ValidateHalConfig(h)
 	if err != nil {
 		t.Errorf("Expected no validation failures, got %s", err)
 	}
@@ -121,7 +121,7 @@ func TestInvalidKubernetesAccount(t *testing.T) {
 			},
 		},
 	}
-	err := validate_hal.ValidateHalConfig(h)
+	err := validate.ValidateHalConfig(h)
 	if err == nil {
 		t.Error("Expected a validation failure, got none")
 	}
